optimization commit
The solution before this commit was way spending way too much time just for getting the name of the country with the certain coordinates. I instead put an arraylist with the US' states in it. From now on if you entered united states to the country input you'll get the output if the place is a state in US.